fix: :sparkles: Rommel - Saber el periodo actual
se agrego un campo en el modelo el cual define el periodo actual, el cual solo debe haber uno

+++ src/school-term/school-term.service.ts
@@ -1,10 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateSchoolTermDto } from './dto/create-school-term.dto';
 import { UpdateSchoolTermDto } from './dto/update-school-term.dto';
 import { InjectModel } from '@nestjs/mongoose';
 import { SchoolTerm } from './model/school-term.entity';
 import { Model } from 'mongoose';
-
 
 @Injectable()
 export class SchoolTermService {
@@ -14,9 +13,15 @@ export class SchoolTermService {
   ) {}
 
   async create(createElectivePeriodDto: CreateSchoolTermDto) {
+    await this.schoolModel.updateMany(
+      { current: true },
+      { $set: { current: false } },
+    );
+
     return await this.schoolModel.create({
       name: createElectivePeriodDto.name,
       description: createElectivePeriodDto.description,
+      current: true,
     });
   }
 
@@ -29,11 +34,17 @@ export class SchoolTermService {
   }
 
   async update(id: string, updateElectivePeriodDto: UpdateSchoolTermDto) {
-    return this.schoolModel.findByIdAndUpdate(
+    await this.schoolModel.updateMany(
+      { current: true },
+      { $set: { current: false } },
+    );
+
+    return await this.schoolModel.findByIdAndUpdate(
       id,
       {
         name: updateElectivePeriodDto.name,
         description: updateElectivePeriodDto.description,
+        current: true,
       },
       {
         new: true,
@@ -42,6 +53,12 @@ export class SchoolTermService {
   }
 
   async remove(id: string) {
+    const findCurrentTerm = await this.schoolModel.findById(id);
+
+    if (findCurrentTerm.current === true) {
+      throw new NotFoundException('Cannot delete current school term');
+    }
+
     return await this.schoolModel.findByIdAndDelete(id);
   }
 }
